+++ realm/realm-library/src/objectServer/java/io/realm/ErrorCode.java
@@ -27,6 +27,8 @@ public enum ErrorCode {
     UNKNOWN(-1),                                // Catch-all
     IO_EXCEPTION(0, Category.RECOVERABLE),      // Some IO error while either contacting the server or reading the response
     JSON_EXCEPTION(1),                          // JSON input could not be parsed correctly
+    CLIENT_RESET(7),                            // Client Reset required.
+
 
     // Realm Object Server errors (100 - 199)
     // Connection level and protocol errors.